+++ contao/languages/de/tl_member.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+$lang = &$GLOBALS['TL_LANG']['tl_member'];
+
+/*
+ * Fields
+ */
+$lang['alias'] = ['Mitgliederalias', 'Der Mitgliederalias ist eine eindeutige Referenz, die anstelle der numerischen Mitglieder-ID aufgerufen werden kann.'];