+++ contao/languages/en/tl_member.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+ * Copyright (c) 2021 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+$lang = &$GLOBALS['TL_LANG']['tl_member'];
+
+/*
+ * Fields
+ */
+$lang['alias'] = ['Member alias', 'The member alias is a unique reference that can be used instead of the member id.'];