+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - 2021-08-02
+- Added: options to add member alias
 - Fixed: added missing doctrine dbal dependency
 
 ## [0.1.1] - 2021-08-02